SDC-42: Fix: 좋아요 수 타입 Long으로 변경

+++ dev-community-backend/src/main/java/com/letsgo/devcommunity/domain/member/service/MemberService.java
@@ -39,7 +39,7 @@ public class MemberService {
                 .orElseThrow(() -> new IllegalArgumentException(NOT_FOUND_MEMBER));
         int followerCount = followRepository.countByToMember(member);
         int followingCount = followRepository.countByFromMember(member);
-        int receivedLikeCount = postLikeRepository.countLikesReceivedByMember(member.getId());
+        Long receivedLikeCount = postLikeRepository.countLikesReceivedByMember(member.getId());
 
         return new MemberProfileResponse(
                 member.getNickname(),
